refactored html to markdown converter
