+++ wiki-fe/src/app/_components/wiki/EditArticle.jsx
@@ -1,14 +1,16 @@
 "use client"
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { getSession } from '@/lib/session';
 import { useRouter } from 'next/navigation';
 import "@/styles/wiki/new-article.css";
 import ErrorMessage from "@/app/_components/auth/ErrorMessage";
-import Showdown from "showdown";
 
 // Importes do Context 
 import { ArticleContext } from '@/app/contexts/articleProvider';
 import { useContext } from 'react';
+
+import useMarkdownToHtml from '@/hooks/markdownToHtml';
+
 
 export default function EditArticle() {
     const router = useRouter();
@@ -25,44 +27,7 @@ export default function EditArticle() {
         setArticleId(article.articleId);
         setText(article.text);
         setTitle(article.title);
-        //renderiza
-        setPreview(article.text);
-
     }, []);
-
-
-    const sd = new Showdown.Converter(
-        {
-            tables: true,
-            tasklists: true,
-            strikethrough: true,
-            emoji: true,
-            simpleLineBreaks: true,
-            openLinksInNewWindow: true,
-            backslashEscapesHTMLTags: true,
-            smoothLivePreview: true,
-            simplifiedAutoLink: true,
-            simpleLineBreaks: true,
-            requireSpaceBeforeHeadingText: true,
-            ghMentions: true,
-            ghMentionsLink: '/user/{u}',
-            ghCodeBlocks: true,
-            emoji: true,
-            underline: true,
-            completeHTMLDocument: true,
-            metadata: true,
-            parseImgDimensions: true,
-            encodeEmails: true,
-            openLinksInNewWindow: true
-        });
-
-    const previewRef = useRef(null);
-    
-    const setPreview = (text) => {
-        if (previewRef.current) {
-            previewRef.current.innerHTML = sd.makeHtml(text);
-        }
-    }
     
     const handleSubmit = async (event) => {
         event.preventDefault();
@@ -125,7 +90,6 @@ export default function EditArticle() {
                             onChange={(e) => 
                                 {
                                     setText(e.target.value)
-                                    setPreview(e.target.value)
                                 }}
                             required
                         />
@@ -134,7 +98,7 @@ export default function EditArticle() {
 
                     <h2>Prévia do artigo</h2>
                     <div className="preview-container">
-                        <div ref={previewRef} className="preview-text"></div>
+                        <div ref={useMarkdownToHtml(text)} className="preview-text"></div>
                     </div>
                 </div>
                 </div>
